refactor robot executer
