+++ tapas-robot-executor/src/main/java/ch/unisg/tapasrobotexecutor/robotexecutor/TapasRobotExecutorApplication.java
@@ -1,4 +1,4 @@
-package ch.unisg.tapasrobotexecutor.tasks;
+package ch.unisg.tapasrobotexecutor.robotexecutor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
